refactor: align namespaces with folder structure (IDE0130)
Move the Cli factory types into Refedle.App.Cli.Factories to match
their directory, and update the source generator's emitted using
to match.
